Refactor view handler logic after fixing chunking bug
