+++ app/src/types/ViewIDs.ts
@@ -1,0 +1,8 @@
+export enum ViewIDs {
+  HOME,
+  DATASETS,
+  EXPERIMENTS,
+  ALGORITHMS,
+  BENCHMARK_CONFIGURATOR,
+  BINARY_METRICS,
+}
